+++ src/app/Providers/MessageRoutingServiceProvider.php
@@ -5,13 +5,11 @@ declare(strict_types=1);
 namespace PostsDMS\Providers;
 
 use Chassis\Framework\Providers\RoutingServiceProvider;
-use PostsDMS\OutboundAdapters\DemoOperationDelete;
-use PostsDMS\OutboundAdapters\DemoOperationDeletedEvents;
-use PostsDMS\OutboundAdapters\DemoOperationGetAsync;
-use PostsDMS\OutboundAdapters\DemoOperationGetSync;
-use PostsDMS\Services\DemoDeleteEventService;
-use PostsDMS\Services\DemoMoreService;
-use PostsDMS\Services\DemoService;
+use PostsDMS\OutboundAdapters\GetAuthorInfoData;
+use PostsDMS\OutboundAdapters\PostCreatedEvent;
+use PostsDMS\OutboundAdapters\PostDeletedEvent;
+use PostsDMS\OutboundAdapters\PostUpdatedEvent;
+use PostsDMS\Services\PostsService;
 
 class MessageRoutingServiceProvider extends RoutingServiceProvider
 {
@@ -19,20 +17,19 @@ class MessageRoutingServiceProvider extends RoutingServiceProvider
      * @var array|string[]
      */
     protected array $inboundRoutes = [
-        'createSomething' => [DemoService::class, 'create'],
-        'getSomething' => [DemoService::class, 'get'],
-        'getSomethingResponse' => [DemoMoreService::class, 'complete'],
-        'deleteSomething' => [DemoService::class, 'delete'],
-        'somethingDeleted' => DemoDeleteEventService::class,
+        'getPost' => [PostsService::class, 'get'],
+        'createPost' => [PostsService::class, 'create'],
+        'updatePost' => [PostsService::class, 'update'],
+        'deletePost' => [PostsService::class, 'delete'],
     ];
 
     /**
      * @var array|string[]
      */
     protected array $outboundRoutes = [
-        'getSomethingSync' => DemoOperationGetSync::class,
-        'getSomethingAsync' => DemoOperationGetAsync::class,
-        'deleteSomething' => DemoOperationDelete::class,
-        'deleteSomethingEvent' => DemoOperationDeletedEvents::class,
+        'getAuthorInfo' => GetAuthorInfoData::class,
+        'postCreated' => PostCreatedEvent::class,
+        'postUpdated' => PostUpdatedEvent::class,
+        'postDeleted' => PostDeletedEvent::class,
     ];
 }